Created a base class for the forward and inverse solver panels to reduce code duplication
